Move a lot of the Bluetooth stuff to a more generic Connection type interface, so that it is easier to add more connection types in the future (such as Wifi Direct)
git-svn-id: https://repo.cs.iastate.edu/coms309/A1/trunk@270 817968c2-256c-45f1-8ff6-e7619683c4bf

+++ Shared/src/cs309/a1/shared/SoundManager.java
@@ -9,40 +9,39 @@ import android.media.MediaPlayer;
 import android.media.SoundPool;
 import android.speech.tts.TextToSpeech;
 import android.util.Log;
-import android.widget.Toast;
 
 public class SoundManager {
-	
+
 	/**
 	 * The Logcat Debug tag
 	 */
 	private static final String TAG = SoundManager.class.getName();
 
-	
+
 	/**
 	 * This has a bunch of sounds that can be played
 	 */
 	private static SoundPool soundpool;
-	
+
 	/**
 	 * This will play music or other sounds
 	 */
 	private static MediaPlayer mediaplayer;
-	
+
 	/**
 	 * This is the code used to access a test sound
 	 */
 	private static int testSound;
-	
+
 	/**
 	 * This is how we will read the names of the players out loud
 	 */
 	private static TextToSpeech tts;
-	
+
 	boolean isTTSInitialized;
-	
+
 	//constants for SoundManager
-	
+
 	/**
 	 * "Hey (player) play a card"
 	 * "Yo (player) you are needed."
@@ -54,38 +53,38 @@ public class SoundManager {
 	public static final int NUM_TURN_STRINGS = 5;
 	public static final String[] playerTurnBeforeName = {"Hey ","Yo ", " ", " ", "Wake up "};
 	public static final String[] playerTurnAfterName = {" play a card!"," you are needed.", " get a job.", ", your turn.", " and smell the waffles."};
-	
-	
-	
+
+
+
 	/**
 	 * @param context
 	 */
 	public SoundManager(Context context){
-		soundpool = new SoundPool(5, AudioManager.STREAM_MUSIC, 100); 
+		soundpool = new SoundPool(5, AudioManager.STREAM_MUSIC, 100);
 		mediaplayer = new MediaPlayer();
 		testSound = soundpool.load(context, R.raw.sound_test, 1);
 		mediaplayer = MediaPlayer.create(context, R.raw.sound_test);
 		MyInitListener mil = new MyInitListener();
 		tts = new TextToSpeech(context, mil);
 		isTTSInitialized = false;
-		
+
 		//TODO this is where you add more sounds
 	}
-	
 
-	
+
+
 	/**
 	 * This will play the test sound
 	 */
 	public static void playTestSound(){
-		soundpool.play(testSound, 1, 1, 1, 0, (float) 1);
+		soundpool.play(testSound, 1, 1, 1, 0, 1);
 	}
-	
+
 	public void playTesttts(){
 		String test = "Hello, this is your text to speech library. Thank you for including me in your project.";
 		speak(test);
 	}
-	
+
 	/**
 	 * This will use TextToSpeech to say the string out loud
 	 * @param words
@@ -96,14 +95,14 @@ public class SoundManager {
 			tts.speak(words, TextToSpeech.QUEUE_FLUSH, null);
 		}
 	}
-	
+
 	/**
 	 * This function will tell a player it is their turn using various strings
 	 * @param name
 	 * the name of the player
 	 */
 	public void sayTurn(String name){
-		 Random r1 = new Random();
+		Random r1 = new Random();
 		int i = Math.abs(r1.nextInt() % NUM_TURN_STRINGS);
 		if (Util.isDebugBuild()) {
 			Log.d(TAG, "Sure " + i);
@@ -111,20 +110,20 @@ public class SoundManager {
 		String words = playerTurnBeforeName[i] + name + playerTurnAfterName[i];
 		tts.speak(words, TextToSpeech.QUEUE_FLUSH, null);
 	}
-	
-	
-	
+
+
+
 	/**
 	 * This will play the theme music
 	 */
 	public static void playMusic(){
-		//probably only do this on Tablet? since it would be really annoying to have 
+		//probably only do this on Tablet? since it would be really annoying to have
 		//multiple songs going on
 		//also this could be used to create sound effects for signaling whose turn it is.
 		mediaplayer.seekTo(0);
 		mediaplayer.start();
 	}
-	
+
 	/**
 	 * This will stop the music from playing
 	 */
@@ -133,7 +132,7 @@ public class SoundManager {
 			mediaplayer.stop();
 		}
 	}
-	
+
 	/**
 	 * This should make all sounds stop playing
 	 */
@@ -144,7 +143,7 @@ public class SoundManager {
 		soundpool.autoPause();
 		tts.stop();
 	}
-	
+
 	private class MyInitListener  implements TextToSpeech.OnInitListener{
 
 		@Override
@@ -157,13 +156,13 @@ public class SoundManager {
 						Log.d(TAG, "Language not available");
 					}
 				} else{
-					isTTSInitialized = true;					
+					isTTSInitialized = true;
 				}
 			} else if (Util.isDebugBuild()) {
 				Log.d(TAG, "Text To Speech did not initialize correctly");
 			}
-	      
+
 		}
-		
+
 	}
 }
